+++ bundle/bundle_builder.go
@@ -8,7 +8,6 @@ import (
 	"bytes"
 	"encoding/binary"
 	"fmt"
-	"sort"
 	"time"
 )
 
@@ -84,8 +83,6 @@ func (bldr *BundleBuilder) Build() (bndl Bundle, err error) {
 	} else {
 		bldr.primary.SetCRCType(bldr.crcType)
 	}
-
-	sort.Sort(canonicalBlockNumberSort(bldr.canonicals))
 
 	bndl, err = NewBundle(bldr.primary, bldr.canonicals)
 	if err == nil {